requirements.txt
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.0rc1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-engineio==4.12.1
python-socketio==5.13.0
pyzbar==0.1.9
qrcode==8.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
